+++ ormar/fields/foreign_key.py
@@ -18,7 +18,6 @@ from typing import (
 import sqlalchemy
 from pydantic import BaseModel, create_model
 from pydantic.typing import ForwardRef, evaluate_forwardref
-from sqlalchemy import UniqueConstraint
 
 import ormar  # noqa I101
 from ormar.exceptions import ModelDefinitionError, RelationshipInstanceError
@@ -160,13 +159,6 @@ def validate_not_allowed_fields(kwargs: Dict) -> None:
         )
 
 
-class UniqueColumns(UniqueConstraint):
-    """
-    Subclass of sqlalchemy.UniqueConstraint.
-    Used to avoid importing anything from sqlalchemy by user.
-    """
-
-
 @dataclass
 class ForeignKeyConstraint:
     """
@@ -242,6 +234,9 @@ def ForeignKey(  # noqa CFQ002
     skip_reverse = kwargs.pop("skip_reverse", False)
     skip_field = kwargs.pop("skip_field", False)
 
+    sql_nullable = kwargs.pop("sql_nullable", None)
+    sql_nullable = nullable if sql_nullable is None else sql_nullable
+
     validate_not_allowed_fields(kwargs)
 
     if to.__class__ == ForwardRef:
@@ -263,6 +258,7 @@ def ForeignKey(  # noqa CFQ002
         alias=name,
         name=kwargs.pop("real_name", None),
         nullable=nullable,
+        sql_nullable=sql_nullable,
         constraints=constraints,
         unique=unique,
         column_type=column_type,