+++ docs/releases.md
@@ -1,3 +1,14 @@
+# 0.10.19
+
+## ✨ Features
+
+* Add support for multi-column non-unique `IndexColumns` in `Meta.constraints` [#307](https://github.com/collerek/ormar/issues/307)
+* Add `sql_nullable` field attribute that allows to set different nullable setting for pydantic model and for underlying sql column [#308](https://github.com/collerek/ormar/issues/308)
+
+## 🐛 Fixes
+
+* Enable caching of relation map to increase performance [#337](https://github.com/collerek/ormar/issues/337)
+
 # 0.10.18
 
 ## 🐛 Fixes